fix: 테이블 구조 및 타입 변경에 따라 코드를 수정한다

+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/common/security/JwtInfo.java
@@ -2,5 +2,5 @@ package com.romanticpipe.reviewcanvas.common.security;
 
 import com.romanticpipe.reviewcanvas.domain.AdminRole;
 
-public record JwtInfo(Long adminId, AdminRole adminRole) {
+public record JwtInfo(Integer adminId, AdminRole adminRole) {
 }
