+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/repository/ReviewDesignRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ReviewDesignRepository extends JpaRepository<ReviewDesign, Long> {
-	Optional<ReviewDesign> findById(Long reviewDesignId);
+public interface ReviewDesignRepository extends JpaRepository<ReviewDesign, Integer> {
+	Optional<ReviewDesign> findById(Integer reviewDesignId);
 
 	List<ReviewDesign> findAllByReviewDesignType(ReviewDesignType reviewDesignType);
 }
